removed VariantInfo and VarintSetInfo cards, all other follow up cards incorporated the code to get basic info from AlleleRegistry API
