+++ components/service/variantinfo/index.tsx
@@ -5,3 +5,4 @@ export * from './alleleSpecificEvidences'
 export * from './xQTLEvidences'
 export * from './externalRecords'
 export * from './hg38GeneAssociations'
+export * from './regulatoryElementsForVariants'
